requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
joblib==1.4.2
numpy==2.1.2
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
threadpoolctl==3.5.0
urllib3==2.2.3
